creo carpeta utils, mejoro notebooks de visualizacion de resultados, agrego nuevamente los high y lows en triple barrier y simulacion, ajusto la logica de trading y la de estrategias
