+++ backbone/backtesting_trader.py
@@ -11,6 +11,7 @@ class BacktestingTrader(ABCTrader):
             self, 
             money: float, 
             trading_strategy, 
+            trading_logic, 
             threshold: float, 
             allowed_days_in_position: int, 
             stop_loss_in_pips: int, 
@@ -25,6 +26,7 @@ class BacktestingTrader(ABCTrader):
         
         super().__init__(
             trading_strategy, 
+            trading_logic, 
             threshold, 
             allowed_days_in_position, 
             stop_loss_in_pips,
