+++ backbone/utils/triple_barrier.py
@@ -45,6 +45,8 @@ def get_daily_volatility(close_prices, span=100):
 
 def apply_triple_barrier(
     close_prices, 
+    high_prices, 
+    low_prices, 
     take_profit_in_pips, 
     stop_loss_in_pips, 
     side,
@@ -71,26 +73,30 @@ def apply_triple_barrier(
         # Evaluar los precios futuros dentro del período máximo de mantenimiento
         for j in range(index + 1, min(index + max_holding_period, len(close_prices))):
             future_close_price = close_prices[j]
+            future_high_price = high_prices[j]
+            future_low_price = low_prices[j]
             
             if side[index] == 1:
 
-                if future_close_price >= upper_barrier_level:
-                    barriers.append((index, 1))  # Etiqueta 1 para toma de ganancias
+                if future_low_price <= lower_barrier_level:
+                    barriers.append((index, 0))  # Etiqueta 0 para stop-loss
                     break
 
-                elif future_close_price <= lower_barrier_level:
-                    barriers.append((index, 0))  # Etiqueta 0 para stop-loss
+                elif future_high_price >= upper_barrier_level:
+                    barriers.append((index, 1))  # Etiqueta 1 para toma de ganancias
                     break
 
             elif side[index] == -1:
                 # Señal de venta: tomar ganancias si se alcanza la barrera inferior
-                if future_close_price <= lower_barrier_level:
+
+                if future_high_price >= upper_barrier_level:
+                    barriers.append((index, 0))  # Etiqueta 0 para stop-loss
+                    break
+
+                elif future_low_price <= lower_barrier_level:
                     barriers.append((index, 1))  # Etiqueta 1 para toma de ganancias
                     break
 
-                elif future_close_price >= upper_barrier_level:
-                    barriers.append((index, 0))  # Etiqueta 0 para stop-loss
-                    break
         else:
             barriers.append((index, 2))  # Etiqueta 2 si no se alcanza ninguna barrera
     
@@ -120,6 +126,8 @@ def apply_triple_barrier(
 
 def triple_barrier_labeling(
         close_prices, 
+        high_prices, 
+        low_prices, 
         take_profit_in_pips, 
         stop_loss_in_pips, 
         side,
@@ -129,6 +137,8 @@ def triple_barrier_labeling(
 
     labels = apply_triple_barrier(
         close_prices,
+        high_prices, 
+        low_prices, 
         take_profit_in_pips, 
         stop_loss_in_pips, 
         side,
@@ -138,12 +148,3 @@ def triple_barrier_labeling(
     
     target = [label for _, label in labels]
     return target
-
-def bbands(close_prices, window, no_of_stdev):
-    rolling_mean = close_prices.ewm(span=window).mean()
-    rolling_std = close_prices.ewm(span=window).std()
-
-    upper_band = rolling_mean + (rolling_std * no_of_stdev)
-    lower_band = rolling_mean - (rolling_std * no_of_stdev)
-
-    return rolling_mean, upper_band, lower_band
