+++ configs/parameters.yml
@@ -1,6 +1,6 @@
 models: [
-  # null, 
-  'random_forest', 
+  null, 
+  # 'random_forest', 
   'gradient_boosting',
   # 'neural_network'
 ]
@@ -10,7 +10,7 @@ train_window: [
   # 5760, # un anio
   # 1920, # cuatro meses
   # 1440, # tres meses
-  # # 1200, # dos meses y medio
+  # 1200, # dos meses y medio
   # 960, # dos meses
   # 720, # mes y medio
   # 480, # un mes
@@ -28,15 +28,15 @@ train_period: [
 stop_loss_in_pips: [
   # 5,
   10,
-  # 15,
+  15,
   # 20,
   # 25
 ]
 
 risk_reward_ratio: [
-  # 2,
-  # 3,
-  4,
+  2,
+  3,
+  # 4,
   # 5,
   # 6, 
   # 7,
@@ -47,7 +47,7 @@ risk_reward_ratio: [
 periods_forward_target: [
   # 4,
   8,
-  # 12,
+  12,
   # 24,
   # 48,
   # 10,
@@ -59,10 +59,13 @@ use_days_in_position: [
 
 use_trailing_stop: [
   True,
-  # False
+  False
 ]
 
 trading_strategy: [
-  'backbone.strategies.ml_strategy', 
-  # 'backbone.strategies.only_strategy', 
+  # 'backbone.strategies.sma_bband_adx_stgy', 
+  # 'backbone.strategies.sma_bband_adx_sell_stgy',
+  'backbone.strategies.bband_sell_stgy',
+  # 'backbone.strategies.rsi_sma_stgy', 
+  # 'backbone.strategies.sma_bband_adx_buy_stgy', 
 ]
