+++ trading.py
@@ -3,7 +3,7 @@ import yaml
 from backbone.machine_learning_agent import MachineLearningAgent
 from backbone.realtime_trader import RealtimeTrader
 from datetime import datetime
-from backbone.utils import load_function
+from backbone.utils.general_purpose import load_function
 from backbone.botardo import Botardo
 from datetime import timedelta
 import os
